Refactor and monomorphize naive recursion search
