+++ core/src/main/scala/eo/analysis/mutualrec/naive/errors.scala
@@ -1,0 +1,10 @@
+package eo.analysis.mutualrec.naive
+
+object errors {
+  case class DuplicatedMethodAttributes(
+    val objectName: String,
+    val methodAttrName: String
+  ) extends Exception(
+    s"""Object ${objectName} defines ${methodAttrName} multiple times."""
+  )
+}
